+++ src/main/java/duke/exception/DukeException.java
@@ -1,4 +1,5 @@
-class DukeException extends Exception {
+package duke.exception;
+public class DukeException extends Exception {
     DukeException(String message) {
         super(message);
     }
