+++ src/main/java/duke/main/Duke.java
@@ -1,3 +1,4 @@
+package duke.main;
 import java.io.*;
 
 public class Duke {
