+++ src/main/java/duke/task/Deadline.java
@@ -1,22 +1,20 @@
+package duke.task;
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-class Event extends Task {
+public class Deadline extends Task {
     private final String taskType;
-    private final LocalDateTime eventStartTime;
-    private final LocalDateTime eventEndTime;
+    private final LocalDateTime taskDeadline;
     DateTimeFormatter dateTimeFormatter1 =
             DateTimeFormatter.ofPattern("MMM dd yyyy HHmm a");
 
-    Event(int taskNumber, boolean taskStatus, String task,
-         LocalDateTime eventStartTime, LocalDateTime eventEndTime, int totalNumOfTasks) {
+    public Deadline(int taskNumber, boolean taskStatus, String task, LocalDateTime taskDeadline, int totalNumOfTasks) {
         super(taskNumber, taskStatus, task, totalNumOfTasks);
-        this.taskType = "[E]";
-        this.eventStartTime = eventStartTime;
-        this.eventEndTime = eventEndTime;
+        this.taskType = "[D]";
+        this.taskDeadline = taskDeadline;
     }
 
     @Override
@@ -24,9 +22,7 @@ class Event extends Task {
         System.out.println("\t____________________________________________________________" +
                 "\n\t Nice! I've marked this task as done:\n" +
                 "\t  " + this.taskType + "[X]" + " " + super.task +
-                " (from: " + this.eventStartTime.format(dateTimeFormatter1)
-                + " to: "
-                + this.eventEndTime.format(dateTimeFormatter1) +
+                " (by: " + this.taskDeadline.format(dateTimeFormatter1) + ")" +
                 "\n\t____________________________________________________________");
     }
 
@@ -35,18 +31,15 @@ class Event extends Task {
         System.out.println("\t____________________________________________________________" +
                 "\n\t OK, I've marked this task as not done yet:\n" +
                 "\t  " + this.taskType + "[ ]" + " " + super.task +
-                " (from: " + this.eventStartTime.format(dateTimeFormatter1)
-                + " to: "
-                + this.eventEndTime.format(dateTimeFormatter1) +
+                " (by: " + this.taskDeadline.format(dateTimeFormatter1) + ")" +
                 "\n\t____________________________________________________________");
     }
 
-    public void printEventTask() {
+    public void printDeadlineTask() {
         System.out.println("\t____________________________________________________________" +
                 "\n\t Got it. I've added this task:" +
-                "\n\t   [E]" + super.getTaskStatus() + " " + super.task +
-                "(from: " + this.eventStartTime.format(dateTimeFormatter1) +
-                " to: " + this.eventEndTime.format(dateTimeFormatter1) + ")" +
+                "\n\t   [D]" + super.getTaskStatus() + " " + super.task +
+                "(by: " + this.taskDeadline.format(dateTimeFormatter1) + ")" +
                 "\n\t Now you have " + super.totalNumOfTasks + " tasks in the list." +
                 "\n\t____________________________________________________________");
     }
@@ -57,12 +50,10 @@ class Event extends Task {
         System.out.println("\t____________________________________________________________" +
                 "\n\t Noted. I've removed this task:" +
                 "\n\t   " + this.taskType +
-                super.getTaskStatus() + " " + super.task + " (from: " +
-                this.eventStartTime.format(dateTimeFormatter1) +
-                " to: " +
-                this.eventEndTime.format(dateTimeFormatter1) +
-                ")" + "\n\t Now you have " +
-                newTotalNumOfTasks + " tasks in the list." +
+                super.getTaskStatus() + " " + super.task + " (by: " +
+                this.taskDeadline.format(dateTimeFormatter1)
+                + ")" +
+                "\n\t Now you have " + newTotalNumOfTasks + " tasks in the list." +
                 "\n\t____________________________________________________________");
     }
 
@@ -72,17 +63,13 @@ class Event extends Task {
     }
 
     @Override
-    public LocalDateTime getEventStartTime() {
-        return this.eventStartTime;
-    }
-
-    @Override
-    public LocalDateTime getEventEndTime() {
-        return this.eventEndTime;
+    public LocalDateTime getDeadline() {
+        return this.taskDeadline;
     }
 
     @Override
     public LocalDate getDate() {
-        return this.eventStartTime.toLocalDate();
+        return this.taskDeadline.toLocalDate();
+
     }
 }
