+++ src/main/java/duke/task/Event.java
@@ -1,19 +1,23 @@
+package duke.task;
 import java.time.LocalDate;
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task {
+public class Event extends Task {
     private final String taskType;
-    private final LocalDateTime taskDeadline;
+    private final LocalDateTime eventStartTime;
+    private final LocalDateTime eventEndTime;
     DateTimeFormatter dateTimeFormatter1 =
             DateTimeFormatter.ofPattern("MMM dd yyyy HHmm a");
 
-    Deadline(int taskNumber, boolean taskStatus, String task, LocalDateTime taskDeadline, int totalNumOfTasks) {
+    public Event(int taskNumber, boolean taskStatus, String task,
+                 LocalDateTime eventStartTime, LocalDateTime eventEndTime, int totalNumOfTasks) {
         super(taskNumber, taskStatus, task, totalNumOfTasks);
-        this.taskType = "[D]";
-        this.taskDeadline = taskDeadline;
+        this.taskType = "[E]";
+        this.eventStartTime = eventStartTime;
+        this.eventEndTime = eventEndTime;
     }
 
     @Override
@@ -21,7 +25,9 @@ public class Deadline extends Task {
         System.out.println("\t____________________________________________________________" +
                 "\n\t Nice! I've marked this task as done:\n" +
                 "\t  " + this.taskType + "[X]" + " " + super.task +
-                " (by: " + this.taskDeadline.format(dateTimeFormatter1) + ")" +
+                " (from: " + this.eventStartTime.format(dateTimeFormatter1)
+                + " to: "
+                + this.eventEndTime.format(dateTimeFormatter1) +
                 "\n\t____________________________________________________________");
     }
 
@@ -30,15 +36,18 @@ public class Deadline extends Task {
         System.out.println("\t____________________________________________________________" +
                 "\n\t OK, I've marked this task as not done yet:\n" +
                 "\t  " + this.taskType + "[ ]" + " " + super.task +
-                " (by: " + this.taskDeadline.format(dateTimeFormatter1) + ")" +
+                " (from: " + this.eventStartTime.format(dateTimeFormatter1)
+                + " to: "
+                + this.eventEndTime.format(dateTimeFormatter1) +
                 "\n\t____________________________________________________________");
     }
 
-    public void printDeadlineTask() {
+    public void printEventTask() {
         System.out.println("\t____________________________________________________________" +
                 "\n\t Got it. I've added this task:" +
-                "\n\t   [D]" + super.getTaskStatus() + " " + super.task +
-                "(by: " + this.taskDeadline.format(dateTimeFormatter1) + ")" +
+                "\n\t   [E]" + super.getTaskStatus() + " " + super.task +
+                "(from: " + this.eventStartTime.format(dateTimeFormatter1) +
+                " to: " + this.eventEndTime.format(dateTimeFormatter1) + ")" +
                 "\n\t Now you have " + super.totalNumOfTasks + " tasks in the list." +
                 "\n\t____________________________________________________________");
     }
@@ -49,10 +58,12 @@ public class Deadline extends Task {
         System.out.println("\t____________________________________________________________" +
                 "\n\t Noted. I've removed this task:" +
                 "\n\t   " + this.taskType +
-                super.getTaskStatus() + " " + super.task + " (by: " +
-                this.taskDeadline.format(dateTimeFormatter1)
-                + ")" +
-                "\n\t Now you have " + newTotalNumOfTasks + " tasks in the list." +
+                super.getTaskStatus() + " " + super.task + " (from: " +
+                this.eventStartTime.format(dateTimeFormatter1) +
+                " to: " +
+                this.eventEndTime.format(dateTimeFormatter1) +
+                ")" + "\n\t Now you have " +
+                newTotalNumOfTasks + " tasks in the list." +
                 "\n\t____________________________________________________________");
     }
 
@@ -62,13 +73,17 @@ public class Deadline extends Task {
     }
 
     @Override
-    public LocalDateTime getDeadline() {
-        return this.taskDeadline;
+    public LocalDateTime getEventStartTime() {
+        return this.eventStartTime;
+    }
+
+    @Override
+    public LocalDateTime getEventEndTime() {
+        return this.eventEndTime;
     }
 
     @Override
     public LocalDate getDate() {
-        return this.taskDeadline.toLocalDate();
-
+        return this.eventStartTime.toLocalDate();
     }
 }
